pyproject.toml
[tool.poetry]
name = "napoletano-web"
version = "0.1.0"
description = ""
authors = ["Baptiste Fontaine <b@ptistefontaine.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.32"
alembic = "^1.7.7"
psycopg2-binary = "^2.9.3"
Flask = "^2.0.3"
gunicorn = "^20.1.0"
Flask-SQLAlchemy = "^2.5.1"
python-slugify = "^6.1.1"

[tool.poetry.dev-dependencies]
mypy = "^0.941"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
